// src/platform/packages/shared/kbn-content-management-utils/src/types.d.ts
export interface Reference {
    type: string;
    id: string;
    name: string;
}
// ...
export type GetResultSO<T extends object = object> = GetResult<T, {
    outcome: 'exactMatch' | 'aliasMatch' | 'conflict';
    aliasTargetId?: string;
    aliasPurpose?: 'savedObjectConversion' | 'savedObjectImport';
}>;
/**
 * Saved object with metadata
 */
export interface SOWithMetadata<Attributes extends object = object> {
    id: string;
    type: string;
    version?: string;
    createdAt?: string;
    updatedAt?: string;
    createdBy?: string;
    updatedBy?: string;
    error?: {
        error: string;
        message: string;
        statusCode: number;
        metadata?: Record<string, unknown>;
    };
    managed?: boolean;
    attributes: Attributes;
    references: Reference[];
    namespaces?: string[];
    originId?: string;
}
export type SOWithMetadataPartial<Attributes extends object = object> = Omit<SOWithMetadata<Attributes>, 'attributes' | 'references'> & {
    attributes: Partial<Attributes>;
    references: Reference[] | undefined;
};
// ...
export interface ContentManagementCrudTypes<ContentType extends string, Attributes extends object, CreateOptions extends object, UpdateOptions extends object, SearchOptions extends object> {
    Attributes: Attributes;
    /**
     * Complete saved object
     */
    Item: SOWithMetadata<Attributes>;
    /**
     * Partial saved object, used as output for update
     */
    PartialItem: SOWithMetadataPartial<Attributes>;
    /**
     * Create options
     */
    CreateOptions: CreateOptions;
    /**
     * Update options
     */
    UpdateOptions: UpdateOptions;
    /**
     * Search options
     */
    SearchOptions: SearchOptions;
    /**
     * Get item params
     */
    GetIn: GetIn<ContentType>;
    /**
     * Get item result
     */
    GetOut: GetResultSO<SOWithMetadata<Attributes>>;
    /**
     * Create item params
     */
    CreateIn: CreateIn<ContentType, Attributes, CreateOptions>;
    /**
     * Create item result
     */
    CreateOut: CreateResult<SOWithMetadata<Attributes>>;
    /**
     * Search item params
     */
    SearchIn: SearchIn<ContentType, SearchOptions>;
    /**
     * Search item result
     */
    SearchOut: SearchResult<SOWithMetadata<Attributes>>;
    /**
     * Update item params
     */
    UpdateIn: UpdateIn<ContentType, Partial<Attributes>, UpdateOptions>;
    /**
     * Update item result
     */
    UpdateOut: UpdateResult<SOWithMetadataPartial<Attributes>>;
    /**
     * Delete item params
     */
    DeleteIn: DeleteIn<ContentType>;
    /**
     * Delete item result
     */
    DeleteOut: DeleteResult;
}